Message Updates, Usage Setting and Bouncing

+++ src/types/usage.rs
@@ -1,7 +1,7 @@
 #[derive(Debug, Clone, Copy)]
 pub struct Usage {
-    pub up: i64,
-    pub down: i64
+    pub up: i128,
+    pub down: i128
 }
 
 // pub type UsageMap = Arc<Mutex<HashMap<String, Usage>>>;
